Buffers are now always Big Endian
And it is no longer possible to configure a buffer to operate in little endian mode.

+++ buffer/src/test/java/io/netty/buffer/api/tests/BufferSendTest.java
@@ -100,7 +100,7 @@ public class BufferSendTest extends BufferTestSupport {
             var send = orig.send();
             verifyInaccessible(orig);
             try (Buffer receive = send.receive()) {
-                assertEquals(42, receive.readInt());
+                assertEquals(42, receive.readLong());
             }
         }
     }
